requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
livereload==2.6.1
MarkupSafe==1.1.1
robot-framework==3.1.2
six==1.12.0
tornado==6.0.3
Werkzeug==0.15.4
